+++ src/components/MenuButton/MenuButton.test.tsx
@@ -58,6 +58,7 @@ describe('MenuButton', () => {
   });
 
   it('should show menu when user clicks on button', async () => {
+    window.HTMLElement.prototype.scrollIntoView = jest.fn();
     render(<Component />);
     fireEvent.click(screen.getByTestId('btn'));
     const menu = await screen.findAllByRole('listbox');
@@ -65,6 +66,7 @@ describe('MenuButton', () => {
   });
 
   it('should show menu when user presses Enter', async () => {
+    window.HTMLElement.prototype.scrollIntoView = jest.fn();
     render(<Component />);
     const btn = await screen.findByTestId('btn');
     fireEvent.keyDown(btn, { key: 'Enter', code: 'Enter', charCode: 13 });
@@ -73,6 +75,7 @@ describe('MenuButton', () => {
   });
 
   it('should show menu when user presses Space', async () => {
+    window.HTMLElement.prototype.scrollIntoView = jest.fn();
     render(<Component />);
     const btn = await screen.findByTestId('btn');
     fireEvent.keyDown(btn, { key: ' ', code: 'Space' });
@@ -81,6 +84,7 @@ describe('MenuButton', () => {
   });
 
   it('should hide menu when user presses Escape', async () => {
+    window.HTMLElement.prototype.scrollIntoView = jest.fn();
     render(<Component />);
     const btn = await screen.findByTestId('btn');
     fireEvent.keyDown(btn, { key: 'Enter', code: 'Enter' });