+++ src/components/OverflowMenu/OverflowMenu.stories.tsx
@@ -87,7 +87,7 @@ const Template1: ComponentStory<typeof OverflowMenu> = (args) => {
     },
   ];
 
-  const [selected, setSelected] = React.useState<string | null>(null);
+  const [selected, setSelected] = React.useState<string | undefined>(undefined);
   const model = useMemo(() => {
     return items.map((item) => ({
       id: item.id,
@@ -161,7 +161,7 @@ const Template2: ComponentStory<typeof OverflowMenu> = (args) => {
     }));
   }, [args.dimension]);
 
-  const [selected, setSelected] = React.useState<string | null>(null);
+  const [selected, setSelected] = React.useState<string | undefined>(undefined);
 
   return (
     <DisplayBlock>
@@ -233,7 +233,7 @@ const Template3: ComponentStory<typeof OverflowMenu> = (args) => {
     return items.map((item) => ({
       id: item.id,
       render: (options: RenderOptionProps) => (
-        <MenuItem dimension={dimensionL} {...options} key={item.id}>
+        <MenuItem dimension={dimensionM} {...options} key={item.id}>
           {item.display}
         </MenuItem>
       ),
@@ -244,14 +244,14 @@ const Template3: ComponentStory<typeof OverflowMenu> = (args) => {
     return items.map((item) => ({
       id: item.id,
       render: (options: RenderOptionProps) => (
-        <MenuItem dimension={dimensionL} {...options} key={item.id}>
+        <MenuItem dimension={dimensionS} {...options} key={item.id}>
           {item.display}
         </MenuItem>
       ),
       disabled: item.disabled,
     }));
   }, []);
-  const [selected, setSelected] = React.useState<string | null>(null);
+  const [selected, setSelected] = React.useState<string | undefined>(undefined);
 
   return (
     <>